comments
'details' part 

i was in a hurry and thought "gamau ribet" later in preprocessing part.  but i didnt realize that i need to add ALL tags possible and i didnt. because of that there are some data i missed. 

next time, if the web has an unorganized tags, just take all text (without defining possible tags) and clean it later so we dont miss any data